refactor(sql-play): split playground into focused modules
Break the single ~440-line component into modules with one responsibility
each, with no behavior change:

- constants.js        seed SQL, default query, tab/column config
- database.js         pure PGlite helpers returning normalized results
- useSqlPlayground.js hook owning all state and the database lifecycle
- ResultView.jsx      presentational rendering (table, status, sub-tabs)
- index.jsx           thin component wiring the hook to the JSX

Also dedupes the repeated loading/error handling that was copy-pasted
across the query/schema/data paths.
